Change code for ipaddr2 address ranges
The ipaddr2 test has to extablish network peering the the Azure IBSm.
The peering has to works for multiple jobs running in parallel.
Each job has to use a different IP range when the peering has to be
created. This commit is reworking an existing and functional
implementation to:
- centralize the calculation of all the addresses and ranges in a single
  place.
- try as much as possible not to use values calculated at library load,
  try to keep th elibrary code as state-less as possible, try avoiding
the read on WORKER_ID in the library.
- try to create an implementation that only use a different ip range for
  each job when the test needs also to create the peering
- convert more lib variables to constants, like `user`
- add more unit tests
- improve the documentation, fix typo, add more test modules POD
  descriptions.